+++ index.html
@@ -9,8 +9,8 @@
   </head>
 
   <body>
-    <div id="root"><h1>SK is here.</h1></div>
+    <div id="root"><h1>Hi S K Reddy, Something Went Wrong</h1></div>
     
-    <script type="module" src="./App.js"></script>
+    <script type="module" src="./src/App.js"></script>
   </body>
 </html>
