+++ src/components/Footer.js
@@ -1,0 +1,9 @@
+const Footer = () => {
+  return (
+    <div className="footer">
+      <h1>This is Footer</h1>
+    </div>
+  );
+};
+
+export default Footer;
