+++ src/components/Header.js
@@ -1,0 +1,18 @@
+import { LOGO_URL } from "../utils/constants";
+const Header = () => {
+  return (
+    <div className="header">
+      <img className="Logo" src={LOGO_URL} alt="Logo" />
+      <div className="nav-items">
+        <ul>
+          <li>Home</li>
+          <li>About Us</li>
+          <li>Contact Us</li>
+          <li>Cart</li>
+        </ul>
+      </div>
+    </div>
+  );
+};
+
+export default Header;
