+++ src/utils/mockData.js
@@ -1,26 +1,3 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-
-const Header = () => {
-  return (
-    <div className="header">
-      <img
-        className="Logo"
-        src="https://lh3.googleusercontent.com/Em7AHf7XBH_RtGfCBVXz9RH8SM_pHkj3xPP-yd3cRguY1_Jc8fmqgx6WxnvGVyPV5xs5gL3HCD0FCuv6Xo4CwoY6ak4"
-        alt="Logo"
-      />
-      <div className="nav-items">
-        <ul>
-          <li>Home</li>
-          <li>About Us</li>
-          <li>Contact Us</li>
-          <li>Cart</li>
-        </ul>
-      </div>
-    </div>
-  );
-};
-
 const resList = [
   {
     type: "restaurant",
@@ -2051,81 +2028,4 @@ const resList = [
   },
 ];
 
-const RestaurentCard = (props) => {
-  const { resData } = props;
-  const {
-    name,
-    cuisines,
-    avgRating,
-    deliveryTime,
-    costForTwo,
-    cloudinaryImageId,
-    aggregatedDiscountInfo,
-    ribbon,
-  } = resData?.data;
-  return (
-    <div className="res-card">
-      <img
-        className="res-img"
-        src={
-          "https://res.cloudinary.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_508,h_320,c_fill/" +
-          cloudinaryImageId
-        }
-        alt="res-img"
-      />
-      {ribbon ? <div className="res-promoted">{ribbon[0]?.type}</div> : null}
-
-      <h2 className="res-name">{name}</h2>
-      <h4 className="res-cuisines">{cuisines.join(", ")}</h4>
-      <div className="res-stars-div">
-        <h4>{avgRating} Star</h4>
-        <h4>{deliveryTime} MINS</h4>
-        <h4>₹{costForTwo / 100} FOR TWO</h4>
-      </div>
-      {aggregatedDiscountInfo?.shortDescriptionList[0]?.meta ? (
-        <p className="res-offers">
-          {aggregatedDiscountInfo?.shortDescriptionList[0]?.meta}
-        </p>
-      ) : null}
-    </div>
-  );
-};
-
-const Body = () => {
-  return (
-    <div className="body">
-      <div className="searchDiv">
-        <input className="input-search" type="text" placeholder="Search for restaurants and food" />
-      </div>
-      <div className="res-container">
-        {resList.map((restaurent) => {
-          return (
-            <RestaurentCard key={restaurent.data.id} resData={restaurent} />
-          );
-        })}
-      </div>
-    </div>
-  );
-};
-
-const Footer = () => {
-  return (
-    <div className="footer">
-      <h1>This is Footer</h1>
-    </div>
-  );
-};
-
-const AppLayout = () => {
-  return (
-    <div className="AppLayout">
-      <Header />
-      <Body />
-      <Footer />
-    </div>
-  );
-};
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-
-root.render(<AppLayout />);
+export default resList;
